+++ src/HttpClient/HttpClientException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Crunz\HttpClient;
+
+use Crunz\Exception\CrunzException;
+
+class HttpClientException extends CrunzException
+{
+}