+++ src/Pinger/PingableException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Crunz\Pinger;
+
+use Crunz\Exception\CrunzException;
+
+class PingableException extends CrunzException
+{
+}